pyproject.toml
[project]
name = "instantclassicmfg"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "spotipy (>=2.25.0,<3.0.0)",
    "typer (>=0.15.2,<0.16.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "ruamel-yaml (>=0.18.10,<0.19.0)"
]

[tool.poetry.scripts]
icm = "icm.cli:cli"

[tool.poetry]
packages = [
    { include = "icm", from = "src" },
]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.4"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110
target-version = ['py310']
include = 'src'
